requirements.txt
# Dependências para o projeto de Análise Preditiva - Versão Compatível

# Análise de dados (versões mais recentes compatíveis)
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Visualização
matplotlib>=3.7.0
seaborn>=0.12.0

# Bancos de dados
pymongo>=4.5.0
psycopg2-binary>=2.9.0

# Utilitários
python-dotenv>=1.0.0
requests>=2.30.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Testes
pytest>=7.4.0
